refactor(web): migrate controller and realtime surfaces to BigDecimal
Updates DemoController, PriceTick, and DemoControllerTest to support
BigDecimal for financial data. Adds JacksonConfig to ensure correct JSON
serialization of BigDecimal as plain numbers.

+++ web/src/main/java/com/tickonomics/web/realtime/PriceTick.java
@@ -1,5 +1,6 @@
 package com.tickonomics.web.realtime;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -9,13 +10,13 @@ import java.util.List;
  */
 public record PriceTick(
     String symbol,
-    double price,
+    BigDecimal price,
     double volume,
     Instant timestamp,
     List<String> conditions,
     String source) {
 
-  public static PriceTick of(String symbol, double price, Instant timestamp) {
+  public static PriceTick of(String symbol, BigDecimal price, Instant timestamp) {
     return new PriceTick(symbol, price, 0.0, timestamp, List.of(), "tickonomics");
   }
 }
